started refactoring
